requirements.txt
# Hypertools Dependencies - Updated for Modern Compatibility
pydata-wrangler>=0.4.0  # Updated to use latest version with Polars backend support
pandas>=1.3.0,<3.0.0    # Broad compatibility, exclude pandas 3.0 breaking changes
numpy>=1.19.0,<2.0.0    # Stable numpy 1.x series

scikit-learn>=1.0.0     # Modern sklearn with consistent API

scipy>=1.7.0            # Compatible with modern numpy/pandas

umap-learn>=0.5.0       # Stable UMAP version

matplotlib>=3.3.0       # Modern matplotlib with better performance
plotly>=5.0.0           # Modern plotly with improved features
seaborn>=0.11.0         # Compatible with matplotlib 3.x

moviepy>=1.0.0          # For animation export
kaleido>=0.2.1          # For static image export from plotly

requests>=2.25.0        # HTTP requests
dill>=0.3.0             # Serialization
six>=1.15.0             # Python 2/3 compatibility (legacy)
configparser>=5.0.0     # Configuration parsing